requirements.txt
flask
flask-cors
google-cloud-aiplatform
vertexai
PyMySQL
python-dotenv
requests
beautifulsoup4
google-cloud-secret-manager
google-auth